+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 * Added `normalize2base` helper client util
 * Added `normalizeMVB` helper client util
 * Added `normalizeRoot` helper client util
+* Added `version` alias information to config
 * Fixed bug preventing user specified `buildEnd` and `transformPageData` from running after theme's
 * Fixed bug preventing `mvb` from correctly setting the `mvbase`
 * Improved `mvb` an `root` link normalization
 * Improved `VPLVersionLink` to only need `version` input
-* Updated to [vitepress@1.5.](https://github.com/vuejs/vitepress/blob/main/CHANGELOG.md)
+* Updated to [vitepress@1.5.0](https://github.com/vuejs/vitepress/blob/main/CHANGELOG.md)
 
 ## v1.1.0-beta.18 - [November 4, 2024](https://github.com/lando/vitepress-theme-default-plus/releases/tag/v1.1.0-beta.18)